Update auto-rds-s3 script (#641)
**Which issue is resolved by this Pull Request:**
Resolves #

**Description of your changes:**
With new credentials option, creating s3 secrets is no longer necessary,
update auto-rds-s3 setup script.

**Testing:**
- [ ] Unit tests pass
- [ ] e2e tests pass
- Details about new tests (If this PR adds a new feature)
- Details about any manual tests performed

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

+++ tests/e2e/utils/rds-s3/auto-rds-s3-cleanup.py
@@ -35,9 +35,10 @@ def delete_s3_bucket(metadata, secrets_manager_client, region):
     else:
         print("Skip deleting S3 bucket...")
 
-    secrets_manager_client.delete_secret(
-        SecretId=metadata["S3"]["secretName"], ForceDeleteWithoutRecovery=True
-    )
+    if metadata["S3"]["secretName"]:
+        secrets_manager_client.delete_secret(
+            SecretId=metadata["S3"]["secretName"], ForceDeleteWithoutRecovery=True
+        )
 
 def check_bucket(bucket_name,s3_client):
     try:
